cleaned up index
